+++ packages/linear-progress/src/index.js
@@ -1,0 +1,1 @@
+export { default, LinearProgressLabel } from "./linear-progress";